+++ src/mimic/src/store/mod.rs
@@ -1,9 +1,23 @@
 pub mod types;
 
 use crate::ic::structures::{btreemap::BTreeMap, DefaultMemory};
+use candid::CandidType;
 use derive_more::{Deref, DerefMut};
+use serde::{Deserialize, Serialize};
 use std::{cell::RefCell, thread::LocalKey};
+use thiserror::Error as ThisError;
 use types::{DataKey, DataValue};
+
+///
+/// StoreError
+/// this also handles errors from macros when looking up stores
+///
+
+#[derive(CandidType, Debug, Serialize, Deserialize, ThisError)]
+pub enum StoreError {
+    #[error("store '{0}' not found")]
+    StoreNotFound(String),
+}
 
 ///
 /// Store